rename package messaging -> integration
