+++ instrumentation/spring/spring-integration-4.1/library/src/main/java/io/opentelemetry/instrumentation/spring/integration/MessageChannelSpanNameExtractor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.spring.messaging;
+package io.opentelemetry.instrumentation.spring.integration;
 
 import io.opentelemetry.instrumentation.api.instrumenter.SpanNameExtractor;
 import org.springframework.integration.channel.AbstractMessageChannel;
